Clean up day 9
I started out thinking I could use matrix arithmetic to calculate distances and then do the moves without stepping, but quickly realized that wouldn't work.
But I still kept a bunch of code around from that, and kind of forced it to work. After some thinking, refactored it to be quite a bit simpler, knowing that any
point can only move at most one space in one direction in any iteration. No more checking for distance, etc. Also realized that part one and part two are
almost identical - the only real difference is that part 1 has only a head and a tail (essentially a rope with 1 segment), but part two has a head, 8 "knots",
and a tail (a rope with 9 segments). This allowed me to make a common function that worked for both, and just need to pass in a count of points.